[nexus] Break up `sagas.rs` into `app/sagas/` directory
