+++ nexus/src/app/instance.rs
@@ -57,7 +57,7 @@ impl super::Nexus {
             )));
         }
 
-        let saga_params = Arc::new(sagas::ParamsInstanceCreate {
+        let saga_params = Arc::new(sagas::instance_create::Params {
             serialized_authn: authn::saga::Serialized::for_opctx(opctx),
             organization_name: organization_name.clone().into(),
             project_name: project_name.clone().into(),
@@ -67,8 +67,8 @@ impl super::Nexus {
 
         let saga_outputs = self
             .execute_saga(
-                Arc::clone(&sagas::SAGA_INSTANCE_CREATE_TEMPLATE),
-                sagas::SAGA_INSTANCE_CREATE_NAME,
+                Arc::clone(&sagas::instance_create::SAGA_TEMPLATE),
+                sagas::instance_create::SAGA_NAME,
                 saga_params,
             )
             .await?;
@@ -198,14 +198,14 @@ impl super::Nexus {
             .await?;
 
         // Kick off the migration saga
-        let saga_params = Arc::new(sagas::ParamsInstanceMigrate {
+        let saga_params = Arc::new(sagas::instance_migrate::Params {
             serialized_authn: authn::saga::Serialized::for_opctx(opctx),
             instance_id: authz_instance.id(),
             migrate_params: params,
         });
         self.execute_saga(
-            Arc::clone(&sagas::SAGA_INSTANCE_MIGRATE_TEMPLATE),
-            sagas::SAGA_INSTANCE_MIGRATE_NAME,
+            Arc::clone(&sagas::instance_migrate::SAGA_TEMPLATE),
+            sagas::instance_migrate::SAGA_NAME,
             saga_params,
         )
         .await?;
